Fixed database arrays
Fixed which arrays will be used to store and display data. There will be
one array containing every reservation, and one array each for every name,
arrivalDate and departureDate. There will also be an array of Integers
which will hold the index of each Reservation that matches a given search.
These indexes will be based off of the array containing every reservation.
The last array will be an array of strings that will contain the names
that will be in the bar on the left. It will be based off of the
reservations that are referenced in the index array; duplicate names will
be removed.