Moved the database label
Moved the database label to the left so that it is easier to read. The
panel it is in now has a boxLayout as it's manager. The other label and
the combo box are now in their own panel so that the combo box doesn't
explode. Things like to explode for some reason.